refactor(data): keep only what is used for country
