--- data-preparation/data/final/multiselect_groups_prod.csv
@@ -1,77 +1,0 @@
-group,country,country_only
-Quickselect Top GHG Emitters in 2022,China,False
-Quickselect Top GHG Emitters in 2022,United States of America,False
-Quickselect Top GHG Emitters in 2022,EU27,False
-Quickselect Top GHG Emitters in 2022,India,False
-Quickselect Top GHG Emitters in 2022,Russian Federation & USSR,False
-Quickselect Top GHG Emitters in 2022,Japan,False
-Quickselect Top GHG Emitters in 2022,Brazil,False
-Quickselect G7,Canada,True
-Quickselect G7,France,True
-Quickselect G7,Germany,True
-Quickselect G7,Italy,True
-Quickselect G7,United Kingdom,True
-Quickselect G7,United States of America,True
-Quickselect G7,Japan,True
-Quickselect G20,Argentina,True
-Quickselect G20,Australia,True
-Quickselect G20,Brazil,True
-Quickselect G20,Canada,True
-Quickselect G20,China,True
-Quickselect G20,France,True
-Quickselect G20,Germany,True
-Quickselect G20,India,True
-Quickselect G20,Indonesia,True
-Quickselect G20,Italy,True
-Quickselect G20,Japan,True
-Quickselect G20,Mexico,True
-Quickselect G20,Russian Federation & USSR,True
-Quickselect G20,Saudi Arabia,True
-Quickselect G20,South Africa,True
-Quickselect G20,South Korea,True
-Quickselect G20,Turkey,True
-Quickselect G20,United Kingdom,True
-Quickselect G20,United States of America,True
-Quickselect BRICS,Brazil,True
-Quickselect BRICS,Russian Federation & USSR,True
-Quickselect BRICS,India,True
-Quickselect BRICS,China,True
-Quickselect BRICS,South Africa,True
-Quickselect Continents,Europe,False
-Quickselect Continents,Asia and Oceania,False
-Quickselect Continents,Africa,False
-Quickselect Continents,Central and South America,False
-Quickselect Continents,North America,False
-Quickselect Continents,Middle East,False
-Quickselect UK-DE-FR-IT,France,True
-Quickselect UK-DE-FR-IT,Italy,True
-Quickselect UK-DE-FR-IT,Germany,True
-Quickselect UK-DE-FR-IT,United Kingdom,True
-Quickselect EU27,Austria,True
-Quickselect EU27,Belgium,True
-Quickselect EU27,Bulgaria,True
-Quickselect EU27,Cyprus,True
-Quickselect EU27,Czechia,True
-Quickselect EU27,Denmark,True
-Quickselect EU27,Estonia,True
-Quickselect EU27,Finland,True
-Quickselect EU27,France,True
-Quickselect EU27,Germany,True
-Quickselect EU27,Greece,True
-Quickselect EU27,Hungary,True
-Quickselect EU27,Ireland,True
-Quickselect EU27,Italy,True
-Quickselect EU27,Latvia,True
-Quickselect EU27,Lithuania,True
-Quickselect EU27,Luxembourg,True
-Quickselect EU27,Malta,True
-Quickselect EU27,Netherlands,True
-Quickselect EU27,Poland,True
-Quickselect EU27,Portugal,True
-Quickselect EU27,Romania,True
-Quickselect EU27,Slovakia,True
-Quickselect EU27,Slovenia,True
-Quickselect EU27,Spain,True
-Quickselect EU27,Sweden,True
-Quickselect EU27,Croatia,True
-Quickselect Top GHG Emitters in 2022,Indonesia,False
