+++ data-preparation/src/countries/main.py
@@ -1,0 +1,14 @@
+import pandas as pd
+from src.countries.multi_selection_country_groups import \
+    process_multi_selection_country_groups
+
+df = pd.read_csv(
+    "src/countries/data/multiselect_groups.csv",
+    sep=",",
+)
+multi_selection_country_groups = process_multi_selection_country_groups(df)
+multi_selection_country_groups.to_csv(
+    "../server/data/country_multiselect_groups.csv",
+    sep=",",
+    index=False,
+)
