--- server/data/COUNTRY_multiselect_groups_prod.csv
@@ -1,77 +1,0 @@
-group,country,country_only
-Quickselect Top GHG Emitters in 2022,China,false
-Quickselect Top GHG Emitters in 2022,United States of America,false
-Quickselect Top GHG Emitters in 2022,EU27,false
-Quickselect Top GHG Emitters in 2022,India,false
-Quickselect Top GHG Emitters in 2022,Russian Federation & USSR,false
-Quickselect Top GHG Emitters in 2022,Japan,false
-Quickselect Top GHG Emitters in 2022,Brazil,false
-Quickselect G7,Canada,true
-Quickselect G7,France,true
-Quickselect G7,Germany,true
-Quickselect G7,Italy,true
-Quickselect G7,United Kingdom,true
-Quickselect G7,United States of America,true
-Quickselect G7,Japan,true
-Quickselect G20,Argentina,true
-Quickselect G20,Australia,true
-Quickselect G20,Brazil,true
-Quickselect G20,Canada,true
-Quickselect G20,China,true
-Quickselect G20,France,true
-Quickselect G20,Germany,true
-Quickselect G20,India,true
-Quickselect G20,Indonesia,true
-Quickselect G20,Italy,true
-Quickselect G20,Japan,true
-Quickselect G20,Mexico,true
-Quickselect G20,Russian Federation & USSR,true
-Quickselect G20,Saudi Arabia,true
-Quickselect G20,South Africa,true
-Quickselect G20,South Korea,true
-Quickselect G20,Turkey,true
-Quickselect G20,United Kingdom,true
-Quickselect G20,United States of America,true
-Quickselect BRICS,Brazil,true
-Quickselect BRICS,Russian Federation & USSR,true
-Quickselect BRICS,India,true
-Quickselect BRICS,China,true
-Quickselect BRICS,South Africa,true
-Quickselect Continents,Europe,false
-Quickselect Continents,Asia and Oceania,false
-Quickselect Continents,Africa,false
-Quickselect Continents,Central and South America,false
-Quickselect Continents,North America,false
-Quickselect Continents,Middle East,false
-Quickselect UK-DE-FR-IT,France,true
-Quickselect UK-DE-FR-IT,Italy,true
-Quickselect UK-DE-FR-IT,Germany,true
-Quickselect UK-DE-FR-IT,United Kingdom,true
-Quickselect EU27,Austria,true
-Quickselect EU27,Belgium,true
-Quickselect EU27,Bulgaria,true
-Quickselect EU27,Cyprus,true
-Quickselect EU27,Czechia,true
-Quickselect EU27,Denmark,true
-Quickselect EU27,Estonia,true
-Quickselect EU27,Finland,true
-Quickselect EU27,France,true
-Quickselect EU27,Germany,true
-Quickselect EU27,Greece,true
-Quickselect EU27,Hungary,true
-Quickselect EU27,Ireland,true
-Quickselect EU27,Italy,true
-Quickselect EU27,Latvia,true
-Quickselect EU27,Lithuania,true
-Quickselect EU27,Luxembourg,true
-Quickselect EU27,Malta,true
-Quickselect EU27,Netherlands,true
-Quickselect EU27,Poland,true
-Quickselect EU27,Portugal,true
-Quickselect EU27,Romania,true
-Quickselect EU27,Slovakia,true
-Quickselect EU27,Slovenia,true
-Quickselect EU27,Spain,true
-Quickselect EU27,Sweden,true
-Quickselect EU27,Croatia,true
-Quickselect Top GHG Emitters in 2022,Indonesia,false
